+++ paft-fundamentals/src/esg.rs
@@ -1,33 +1,34 @@
 //! ESG scores and involvement types.
 
+use paft_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 
 #[cfg(feature = "dataframe")]
 use df_derive::ToDataFrame;
 
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Default)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
 #[cfg_attr(feature = "dataframe", derive(ToDataFrame))]
 /// ESG involvement details for controversial activities or sectors.
 pub struct EsgInvolvement {
     /// Involvement category.
     pub category: String,
     /// Provider-specific involvement score or flag.
-    pub score: Option<f64>,
+    pub score: Option<Decimal>,
 }
 
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Default)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
 #[cfg_attr(feature = "dataframe", derive(ToDataFrame))]
 /// ESG scores summary.
 pub struct EsgScores {
     /// Environmental score.
-    pub environmental: Option<f64>,
+    pub environmental: Option<Decimal>,
     /// Social score.
-    pub social: Option<f64>,
+    pub social: Option<Decimal>,
     /// Governance score.
-    pub governance: Option<f64>,
+    pub governance: Option<Decimal>,
 }
 
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize, Default)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
 /// ESG summary including scores and involvement details.
 pub struct EsgSummary {
     /// Optional aggregate scores.
